Navbar added.
